+++ docs/README.md
@@ -11,17 +11,11 @@
 [네이버 코딩 컨벤션](https://naver.github.io/hackday-conventions-java/#_intellij)
 
 ## 기능목록
-- 예외처리 환경 구축
 
+- 예외처리 환경 구축
 - 장소 추천 기능
     - 지역과 키워드를 받아 장소를 추천.
     - 예외처리
         - 외부 API 통신 예외처리
-<<<<<<< Updated upstream
-            
-    
+        - 페이지 사이즈 범위 예외 체크
 
-=======
-        
-            
->>>>>>> Stashed changes
