--- src/main/java/ddd/caffeine/ratrip/RecommendPlaceUseCase.java
@@ -1,7 +1,0 @@
-package ddd.caffeine.ratrip;
-
-import ddd.caffeine.ratrip.service.KakaoFeignResponseDto;
-
-public interface RecommendPlaceUseCase {
-	KakaoFeignResponseDto recommendPlace(String region, String keyword);
-}
