+++ src/main/java/ddd/caffeine/ratrip/exception/domain/KakaoFeignException.java
@@ -1,6 +1,7 @@
 package ddd.caffeine.ratrip.exception.domain;
 
 import ddd.caffeine.ratrip.exception.BaseException;
+import ddd.caffeine.ratrip.exception.ExceptionInformation;
 import lombok.Getter;
 
 @Getter
@@ -8,5 +9,9 @@ public class KakaoFeignException extends BaseException {
 
 	public KakaoFeignException(int status, String errorCode, String message) {
 		super(status, errorCode, message);
+	}
+
+	public KakaoFeignException(ExceptionInformation information) {
+		super(information);
 	}
 }
