--- src/main/java/ddd/caffeine/ratrip/service/KakaoFeignResponseDto.java
@@ -1,13 +1,0 @@
-package ddd.caffeine.ratrip.service;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@NoArgsConstructor
-public class KakaoFeignResponseDto {
-    List<KakaoFeignModel> documents;
-    KakaoFeignMetaData meta;
-}
