+++ src/main/java/ddd/caffeine/ratrip/service/RecommendPlaceUseCase.java
@@ -1,0 +1,7 @@
+package ddd.caffeine.ratrip.service;
+
+import ddd.caffeine.ratrip.service.feign.model.KakaoFeignResponseDto;
+
+public interface RecommendPlaceUseCase {
+	KakaoFeignResponseDto recommendPlace(String region, String keyword, int page);
+}
